move container files to components dir
