+++ src/components/routes/Home/Home.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Divider, Header, Icon, Step } from 'semantic-ui-react';
 
-import { NewRoomButton, TopBar } from '../../../components/TopBar';
+import { NewRoomButton, TopBar } from '../../TopBar';
 import styles from './Home.module.css';
 
 export const Home: React.FC = () => {
